feat: Change i64 to usize

+++ src/token.rs
@@ -3,5 +3,5 @@ use crate::token_type::TokenType;
 pub struct Token {
     token_type: TokenType,
     lexeme: String,
-    line: i64,
+    line: usize,
 }
